CassandraClientPool -> CassandraClientPoolImpl
