+++ atlasdb-cassandra-integration-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraClientPoolIntegrationTest.java
@@ -55,8 +55,6 @@ import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceConfigManager;
 import com.palantir.atlasdb.cassandra.ImmutableCassandraKeyValueServiceConfig;
 import com.palantir.atlasdb.containers.CassandraContainer;
 import com.palantir.atlasdb.containers.Containers;
-import com.palantir.atlasdb.keyvalue.cassandra.CassandraClientPool.LightweightOppToken;
-import com.palantir.atlasdb.keyvalue.cassandra.CassandraClientPool.WeightedHosts;
 import com.palantir.common.base.FunctionCheckedException;
 
 public class CassandraClientPoolIntegrationTest {
@@ -68,7 +66,7 @@ public class CassandraClientPoolIntegrationTest {
     private CassandraKeyValueService kv = CassandraKeyValueService.create(
             CassandraKeyValueServiceConfigManager.createSimpleManager(CassandraContainer.KVS_CONFIG),
             CassandraContainer.LEADER_CONFIG);
-    private CassandraClientPool clientPool = kv.getClientPool();
+    private CassandraClientPoolImpl clientPool = (CassandraClientPoolImpl) kv.getClientPool();
 
     @Before
     public void setUp() {
@@ -84,10 +82,12 @@ public class CassandraClientPoolIntegrationTest {
     // Pretty legit test if run manually or if we go back to multi-node tests
     @Test
     public void testTokenMapping() {
-        Map<Range<LightweightOppToken>, List<InetSocketAddress>> mapOfRanges = clientPool.tokenMap.asMapOfRanges();
+        Map<Range<CassandraClientPoolImpl.LightweightOppToken>, List<InetSocketAddress>> mapOfRanges =
+                clientPool.tokenMap.asMapOfRanges();
 
-        for (Entry<Range<LightweightOppToken>, List<InetSocketAddress>> entry : mapOfRanges.entrySet()) {
-            Range<LightweightOppToken> tokenRange = entry.getKey();
+        for (Entry<Range<CassandraClientPoolImpl.LightweightOppToken>, List<InetSocketAddress>> entry :
+                mapOfRanges.entrySet()) {
+            Range<CassandraClientPoolImpl.LightweightOppToken> tokenRange = entry.getKey();
             List<InetSocketAddress> hosts = entry.getValue();
 
             clientPool.getRandomHostForKey("A".getBytes(StandardCharsets.UTF_8));
@@ -180,25 +180,26 @@ public class CassandraClientPoolIntegrationTest {
 
     @Test
     public void testIsConnectionException() {
-        assertFalse(CassandraClientPool.isConnectionException(new TimedOutException()));
-        assertFalse(CassandraClientPool.isConnectionException(new TTransportException()));
-        assertTrue(CassandraClientPool.isConnectionException(new TTransportException(new SocketTimeoutException())));
+        assertFalse(CassandraClientPoolImpl.isConnectionException(new TimedOutException()));
+        assertFalse(CassandraClientPoolImpl.isConnectionException(new TTransportException()));
+        assertTrue(CassandraClientPoolImpl.isConnectionException(new TTransportException(
+                new SocketTimeoutException())));
     }
 
     @Test
     public void testIsRetriableException() {
-        assertTrue(CassandraClientPool.isRetriableException(new TimedOutException()));
-        assertTrue(CassandraClientPool.isRetriableException(new TTransportException()));
-        assertTrue(CassandraClientPool.isRetriableException(new TTransportException(new SocketTimeoutException())));
+        assertTrue(CassandraClientPoolImpl.isRetriableException(new TimedOutException()));
+        assertTrue(CassandraClientPoolImpl.isRetriableException(new TTransportException()));
+        assertTrue(CassandraClientPoolImpl.isRetriableException(new TTransportException(new SocketTimeoutException())));
     }
 
     @Test
     public void testIsRetriableWithBackoffException() {
-        assertTrue(CassandraClientPool.isRetriableWithBackoffException(new NoSuchElementException()));
-        assertTrue(CassandraClientPool.isRetriableWithBackoffException(new UnavailableException()));
-        assertTrue(CassandraClientPool.isRetriableWithBackoffException(
+        assertTrue(CassandraClientPoolImpl.isRetriableWithBackoffException(new NoSuchElementException()));
+        assertTrue(CassandraClientPoolImpl.isRetriableWithBackoffException(new UnavailableException()));
+        assertTrue(CassandraClientPoolImpl.isRetriableWithBackoffException(
                 new TTransportException(new SocketTimeoutException())));
-        assertTrue(CassandraClientPool.isRetriableWithBackoffException(
+        assertTrue(CassandraClientPoolImpl.isRetriableWithBackoffException(
                 new TTransportException(new UnavailableException())));
     }
 
@@ -209,7 +210,7 @@ public class CassandraClientPoolIntegrationTest {
                 new InetSocketAddress(1), createMockClientPoolingContainerWithUtilization(10),
                 new InetSocketAddress(2), createMockClientPoolingContainerWithUtilization(10));
 
-        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPool.WeightedHosts.create(pools).hosts;
+        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPoolImpl.WeightedHosts.create(pools).hosts;
 
         int expectedWeight = result.firstEntry().getKey();
         int prevKey = 0;
@@ -228,7 +229,7 @@ public class CassandraClientPoolIntegrationTest {
                 new InetSocketAddress(1), createMockClientPoolingContainerWithUtilization(10),
                 lowActivityHost, createMockClientPoolingContainerWithUtilization(0));
 
-        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPool.WeightedHosts.create(pools).hosts;
+        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPoolImpl.WeightedHosts.create(pools).hosts;
 
         int largestWeight = result.firstEntry().getKey();
         InetSocketAddress hostWithLargestWeight = result.firstEntry().getValue();
@@ -252,7 +253,7 @@ public class CassandraClientPoolIntegrationTest {
                 new InetSocketAddress(1), createMockClientPoolingContainerWithUtilization(5),
                 highActivityHost, createMockClientPoolingContainerWithUtilization(20));
 
-        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPool.WeightedHosts.create(pools).hosts;
+        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPoolImpl.WeightedHosts.create(pools).hosts;
 
         int smallestWeight = result.firstEntry().getKey();
         InetSocketAddress hostWithSmallestWeight = result.firstEntry().getValue();
@@ -275,7 +276,7 @@ public class CassandraClientPoolIntegrationTest {
                 new InetSocketAddress(1), createMockClientPoolingContainerWithUtilization(10),
                 new InetSocketAddress(2), createMockClientPoolingContainerWithUtilization(15));
 
-        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPool.WeightedHosts.create(pools).hosts;
+        NavigableMap<Integer, InetSocketAddress> result = CassandraClientPoolImpl.WeightedHosts.create(pools).hosts;
 
         int prevKey = 0;
         for (Map.Entry<Integer, InetSocketAddress> entry : result.entrySet()) {
@@ -292,7 +293,7 @@ public class CassandraClientPoolIntegrationTest {
                 new InetSocketAddress(0), createMockClientPoolingContainerWithUtilization(5),
                 new InetSocketAddress(1), createMockClientPoolingContainerWithUtilization(10),
                 new InetSocketAddress(2), createMockClientPoolingContainerWithUtilization(15));
-        WeightedHosts weightedHosts = WeightedHosts.create(pools);
+        CassandraClientPoolImpl.WeightedHosts weightedHosts = CassandraClientPoolImpl.WeightedHosts.create(pools);
         Map<InetSocketAddress, Integer> hostsToWeight = new HashMap<>();
         int prevKey = 0;
         for (Map.Entry<Integer, InetSocketAddress> entry : weightedHosts.hosts.entrySet()) {
