+++ atlasdb-cassandra-integration-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/SchemaMutationLockIntegrationTest.java
@@ -98,7 +98,7 @@ public class SchemaMutationLockIntegrationTest {
                 CassandraKeyValueServiceConfigManager.createSimpleManager(quickTimeoutConfig);
         TracingQueryRunner queryRunner = new TracingQueryRunner(log, TracingPrefsConfig.create());
         writeConsistency = ConsistencyLevel.EACH_QUORUM;
-        clientPool = new CassandraClientPool(simpleManager.getConfig());
+        clientPool = CassandraClientPoolImpl.create(simpleManager.getConfig());
         lockTable = new UniqueSchemaMutationLockTable(
                 new SchemaMutationLockTables(clientPool, quickTimeoutConfig),
                 LockLeader.I_AM_THE_LOCK_LEADER);
